Moved the diffrence classes to their corresponding packages (#3)
