+++ src/main/java/VotingApp/user/User.java
@@ -1,5 +1,6 @@
-package VotingApp;
+package VotingApp.user;
 
+import VotingApp.poll.Poll;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
